[PLAT-148511] Run ref-checks when there are no type errors (#271)
Previously, the presentation compiler would stop at typer and miss
diagnostics emitted by later phases. In particular, the refhcecks phase
is responsible for overriding and inheritance checks, which issues very
useful errors like "missing implementation" for abstract members.

Now, we run an additional phase, refhecks, when there are no typer
errors in a file. There is some incidental complexity around compiler
bookkeeping and hiding of expected errors like "macro not expanded".
Another gotcha is that refchecks applies transformations to the tree,
such as replacing case class apply calls with `new` instantiations
where possible. We save and restore the original typed AST as to not
break existing code, like find references, that depended on the old 
tree shape.

Will add a feature flag and setting to enable/disable this feature,
which is supposedly relatively expensive to compute.