[PLAT-148990] Fix errors involving case classes loaded from source (#299)
The presentation compiler sometimes fails to find the synthetic
companion object of a case class, in particular after a restart. This
can manifest through spurious errors that go away on their own or on
first edit, but otherwise, as in the case of S3UploaderSuite/Config,
they can be persistent.

To fix this, we inline and adapt the Scala BrowsingLoaders component,
which is responsible for creating and entering symbols loaded from the
sourcepath. In this case, if a class has a CASE modifier we immediately
enter a companion class for it. Triplequote Hydra hit the same issue
years back, and we had the same fix.