Fix crash in BrowsingLoaders (#301)
Set an explicit unit when parsing sources for loading a new package. The
current logic is using the global unit, which is the one triggering the
sourcepath-based completer, but in doing so, it's parsing other files.
Deprecation warnings in the scanner (in particular about the use of
unicode arrow, =>) are then emitted and positioned at an offset that's
not matching the global unit, causing crashes.

I couldn't reproduce it in a reliable way in order to have a test, but
it's a fairly clear fix based on stack traces.