Improve ranking of Protobuf search results (#310)
Previously, we returned workspace/symbol results in no particular order.
This was annoying when searching for proto symbols in the universe repo,
where we have this convention for protobuf files where we create a full
copy of the file in a `version_history/vX.proto` when we edit it, and
only ever edit the `latest.proto` file.

Now, we have a very basic ranking heuristic for workspace/symbols where
we sort Proto symbols additionally by the URI so that `latest.proto`
results rank higher than `version_history/vX.proto` results. It's not an
ideal solution, but it solves an acute problem and it shouldn't
negatively influence other results for now. Eventually, I'd love to have
better general ranking of symbol results.

## Test Plan

I manually tested this in universe for the query `EntityAppInstance`
which has hundreds of results.

<img width="720" height="612" alt="image"
src="https://github.com/user-attachments/assets/fdbf1e70-f90c-4481-b1be-d0d0a73be853"
/>

I considered adding tests, but we don't have protobuf indexing available
in the current workspace/symbol tests
so I figured we can let it fly for now.