Run LSP command for file-sync in the background (#312)
Previously, the file-sync LSP command awaited until the Bazel sync
completed AND we had reindexed the new workspace. This was convenient
for testing but it made LSP clients like Helix report an error because
these steps take longer than 30s to run. Now, we return immediately
from the command outside of testing.

<!-- jj-stack-begin -->
- → **#312** ←
[[Commit](https://github.com/REDACTED_ORG/metals/pull/312/commits/9d4d455e48802eb6342fb15dff51522726260150)]
<!-- jj-stack-end -->