Override more javac services behind feature flags (#311)
While working on the upcoming Turbine classpath mode, I tested Metals in
many large repos and triggered a ton of find references that tripped up
out existing javac integration. This PR introduces several critical
fixes to stabilize the java support so that it doesn't crash and fail
fast as frequently when encountering surprising code. These changes
require overriding more internal parts of the java compiler so they're
gated behind individual feature flags making it easy to rollback these
changes if necessary.

## Why separate feature flags?

I debating having just one feature flag. The main reason I want one FF
per service is that we may want to override a fifth/sixth service in the
future, and I don't want to use a single monolithic setting for all of
them. It would be a regression if we're overriding four services,
which fixes many bugs, and we then rollback all of them because a fifth
service is causing problems.

<!-- jj-stack-begin -->
- → **#311** ←
[[Commit](https://github.com/REDACTED_ORG/metals/pull/311/commits/274dfb74769cd05a32f83597083597202cd7d290)]
- #307
[[Commit](https://github.com/REDACTED_ORG/metals/pull/307/commits/1727b0087fee80d1745b5e2cb28195492cfcfb12)]
- #264
[[Commit](https://github.com/REDACTED_ORG/metals/pull/264/commits/8dcab342bc1040817e20670743edec4dd597d2f9)]
<!-- jj-stack-end -->