Refactoring: Introduce `JavaFileManagerFactory` (#307)
This is the last major refactoring in preparation for Turbine.

Previously, the mtags-java module used the SemanticdbFileManager to list
sources in a given package, and then implemented a JavaFileManager class
inside mtags-java using that class. This architecture has been a major
pain point as I worked on adding the Turbine classpath mode because it
requires encoding a ton of APIs from Metals via mtags-interfaces. All
that was unnecessary ceremony becausse mtags-interfaces can just use
javax.tools.JavaFileManager directly instead, and we can implement the
file manager on the Metals side.

This PR is a pure refactoring moving the existing java sourcepath
support over to the new JavaFileManagerFactory API, and it introduces a
new config option javaSymbolLoader that only has an implementation for
the `java-sourcepath` value (`turbine-classpath`). Once this refactoring
lands, the actual turbine-classpath mode is going to be a clean diff
since it's just a sibling implementation of
`JavacSourcepathFileManager`.