Add Java Turbine Classpath Mode (#264)
Previously, we only supported one `javaSymbolLoader: "javac-sourcepath"`
setting, which loaded 1st-party symbols through the javac `-sourcepath`
mode. This was problematic because javac loads the sourcepath eagerly,
which could have noticeable performance overhead in large codebases
where a single file could potentially import over a million lines of
code on the sourcepath.

Now, there's a new `javaSymbolLoader: "turbine-classpath"` mode, which
is makes javac load 1st-party symbols through the `-classpath` where the
classpath is provided by Turbine, a Java header compiler (mostly used by
Bazel). Header compilation is very fast, my benchmarks show it processes
upwards of one million lines of Java code per *second*. This fast
performance means that we can build a classpath of even large Java
codebases on every startup. The tricky bit is keeping the classpath
up-to-date with new changes, and we solve that in this PR by keeping
track of changed paths and adding them to the sourcepath until we
complete the next round of header compilation (which is scheduled to run
every ~minute while you're editing files).

## Test Plan

See the added JavaPCDiagnosticsTrickySuite and
JavaPCTurbineSourcepathSuite. These add test coverage for a large number
of situations where Turbine classpath needs to fallback to using
SOURCE_PATH for recently edited/removed files. We never added these
kinds of tests for javac-sourcepath because it sort of handles them
automatically by construction, while we need really tricky code to
handle the same with turbine-classpath.

I also manually tested this PR and confirmed that it picks up symbol
changes as you change files, and it reports errors when you remove
files.

- Add/change/remove symbols in one file and consume them in another file
- Add a new file and confirm it's picked up
- Remove a file and confirm it results in errors in files that depend on
  that file

## Rollout Plan

This functionality is feature flagged and the default will remain
`javac-sourcepath` so we can test this functionality in canary a bit
more before rolling it out more widely.